+++ .github/workflows/build.yaml
@@ -60,3 +60,9 @@ jobs:
 
     - name: run all rust unittests
       run: cargo test --all
+
+    - name: Run llvm-cov
+      run: cargo llvm-cov --codecov --output-path codecov.json --ignore-filename-regex tonic-h3-tests.*
+
+    - name: Show cov summary
+      run: cargo llvm-cov report --summary-only --ignore-filename-regex tonic-h3-tests.*
